+++ synergy/plot/constants.py
@@ -1,6 +1,0 @@
-"""Some useful constants"""
-
-# Colormap defaults to keep plots consistent
-CMAP_E = "YlGnBu"
-CMAP_RESIDUAL = "RdBu"
-CMAP_SYNERGY = "PRGn"